Add /etc/default/snmpd
Rename OPTIONS to SNMPD_OPTIONS and move them to /etc/default/snmpd.

Additionally, pull pidfile out of it and put it in separate variable
SNMPD_PIDFILE so we don't need to hardcode it in condrestart.
Ditto for SNMPD_CONFIG (checked in start()). While at it, move
-c $SNMPD_CONFIG before $SNMPD_OPTIONS, which makes it possible for
administrator to start snmpd with another config file by putting
-c /some/other/file in $SNMPD_OPTIONS.

Expose only SNMPD_OPTIONS in /etc/default/snmpd, commented out, with
value denoting the defaults used.

The only change compared to stock Slackware is that command used to
start snmpd is echoed before, not after, calling it.

Signed-off-by: Jakub Jankowski <shasta@toxcorp.com>